Model/requirements.txt
torch
faiss-cpu
pandas
numpy
pillow
transformers
sentence-transformers
google-generativeai
gradio
pyarrow